refactor: group console driver package into APIs
