+++ libs/ngworker/lumberjack/console-driver/src/lib/configuration/console-driver-config.token.ts
