+++ libs/ngworker/lumberjack/console-driver/src/lib/log-driver/console.driver.spec.ts
@@ -4,9 +4,10 @@ import { SpyConsole, SpyConsoleModule } from '@internal/console-driver/test-util
 import { resolveDependency } from '@internal/test-util';
 import { LogDriver, logDriverToken, LumberjackLevel } from '@ngworker/lumberjack';
 
-import { ConsoleDriverModule } from './console-driver.module';
+import { ConsoleDriverModule } from '../configuration/console-driver.module';
+import { lumberjackConsoleToken } from '../console/lumberjack-console.token';
+
 import { ConsoleDriver } from './console.driver';
-import { lumberjackConsoleToken } from './lumberjack-console.token';
 
 describe(ConsoleDriver.name, () => {
   beforeEach(() => {
